+++ src/project_info.h
@@ -3,7 +3,7 @@
  * Programmer: Aaron Morris
  * Company:    A2ThreeD
  * Version:    1.3.0
- * Date:       2026-08-10
+ * Date:       2026-08-11
  *
  * Copyright (c) 2026 Aaron Morris / A2ThreeD.
  *
@@ -34,7 +34,7 @@
 #define PROJECT_VERSION_MINOR 3
 #define PROJECT_VERSION_PATCH 0
 #define PROJECT_VERSION_STRING "1.3.0"
-#define PROJECT_RELEASE_DATE "2026-08-10"
+#define PROJECT_RELEASE_DATE "2026-08-11"
 
 // License and release summary reported by the Pico binary metadata.
 #define PROJECT_LICENSE "PolyForm-Noncommercial-1.0.0"